call the accessor, don't write directly to @attributes
I'm not quite sure of the ramifications of this - it may need to be done on the attributes defined as virtual only.  By calling send - we make sure that those methods are called, as well as any methods overriding standard attribute accessors.  This seems like the smart idea, and I probably should check upstream and ask about why attributes were being written directly to @attributes and not through the accessors.